Initial commit
These scripts aim to make the generation of POSCAR files for 3-order
calculations, as well as the harvesting of results to create a
FORCE_CONSTANTS_3RD file, easier for the user. They are based on a
combination of our previous Fortran code plus a thin wrapper for some of
the features that spglib offers.